+++ index.js
@@ -11,46 +11,20 @@
 
   'use strict';
 
-  var failed = false;
-
-  function onFail(message) {
+  function fail(message) {
     // start debugger only once
-    if (!failed) {
-      /*
-        DEBUG HINT:
-        if you are reading this, chances are that there is a bug in your system
-        see the Call Stack to find out what's wrong..
-      */
+    if (!fail.failed) {
       /*jshint debug: true*/
       debugger;
     }
-    failed = true;
-    throw new Error(message);
+    fail.failed = true;
+    throw new TypeError(message);
   }
 
-  var options = {
-    onFail: onFail
-  };
-
-  function fail(message) {
-    /*
-      DEBUG HINT:
-      if you are reading this, chances are that there is a bug in your system
-      see the Call Stack to find out what's wrong..
-    */
-    options.onFail(message);
-  }
-
-  function assert(guard) {
+  function assert(guard, message) {
     if (guard !== true) {
-      var args = slice.call(arguments, 1);
-      var message = args[0] ? format.apply(null, args) : 'assert failed';
-      /*
-        DEBUG HINT:
-        if you are reading this, chances are that there is a bug in your system
-        see the Call Stack to find out what's wrong..
-      */
-      fail(message);
+      message = message ? format.apply(null, slice.call(arguments, 1)) : 'assert failed';
+      exports.fail(message);
     }
   }
 
@@ -61,13 +35,11 @@
   var slice = Array.prototype.slice;
 
   function mixin(target, source, overwrite) {
-    if (Nil.is(source)) {
-      return target;
-    }
+    if (Nil.is(source)) { return target; }
     for (var k in source) {
       if (source.hasOwnProperty(k)) {
         if (overwrite !== true) {
-          assert(!target.hasOwnProperty(k), 'cannot overwrite property %s', k);
+          assert(!target.hasOwnProperty(k), 'Cannot overwrite property %s', k);
         }
         target[k] = source[k];
       }
@@ -96,17 +68,19 @@
     return str;
   }
 
+  function getFunctionName(f) {
+    assert(typeof f === 'function', 'Invalid argument `f` = `%s` supplied to `getFunctionName()`', f);
+    return f.displayName || f.name || format('<function%s>', f.length);
+  }
+
   function replacer(key, value) {
-    if (typeof value === 'function') {
-      return format('Func', value.name);
-    }
-    return value;
+    return Func.is(value) ? getFunctionName(value) : value;
   }
 
   format.formatters = {
-    s: function formatString(x) { return String(x); },
-    j: function formatJSON(x) {
-      try {
+    s: function (x) { return String(x); },
+    j: function (x) {
+      try { // handle circular references
         return JSON.stringify(x, replacer);
       } catch (e) {
         return String(x);
@@ -114,18 +88,13 @@
     }
   };
 
-  function getName(type) {
-    assert(Type.is(type), 'Invalid argument `type` of value `%j` supplied to `getName()`, expected a type.', type);
+  function getTypeName(type) {
+    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `getTypeName()`', type);
     return type.meta.name;
   }
 
-  function getKind(type) {
-    assert(Type.is(type), 'Invalid argument `type` of value `%j` supplied to `geKind()`, expected a type.', type);
-    return type.meta.kind;
-  }
-
   function blockNew(x, type) {
-    assert(!(x instanceof type), 'Operator `new` is forbidden for `%s`', getName(type));
+    assert(!(x instanceof type), 'Operator `new` is forbidden for type `%s`', getTypeName(type));
   }
 
   function shallowCopy(x) {
@@ -149,16 +118,16 @@
   }
 
   update.commands = {
-    '$apply': function $apply(f, value) {
+    '$apply': function (f, value) {
       assert(Func.is(f));
       return f(value);
     },
-    '$push': function $push(elements, arr) {
+    '$push': function (elements, arr) {
       assert(Arr.is(elements));
       assert(Arr.is(arr));
       return arr.concat(elements);
     },
-    '$remove': function $remove(keys, obj) {
+    '$remove': function (keys, obj) {
       assert(Arr.is(keys));
       assert(Obj.is(obj));
       for (var i = 0, len = keys.length ; i < len ; i++ ) {
@@ -166,18 +135,18 @@
       }
       return obj;
     },
-    '$set': function $set(value) {
+    '$set': function (value) {
       return value;
     },
-    '$splice': function $splice(splices, arr) {
+    '$splice': function (splices, arr) {
       assert(list(Arr).is(splices));
       assert(Arr.is(arr));
-      return splices.reduce(function reducer(acc, splice) {
+      return splices.reduce(function (acc, splice) {
         acc.splice.apply(acc, splice);
         return acc;
       }, arr);
     },
-    '$swap': function $swap(config, arr) {
+    '$swap': function (config, arr) {
       assert(Obj.is(config));
       assert(Num.is(config.from));
       assert(Num.is(config.to));
@@ -187,7 +156,7 @@
       arr[config.from] = element;
       return arr;
     },
-    '$unshift': function $unshift(elements, arr) {
+    '$unshift': function (elements, arr) {
       assert(Arr.is(elements));
       assert(Arr.is(arr));
       return elements.concat(arr);
@@ -198,26 +167,17 @@
   };
 
   //
-  // irriducibles
+  // irreducibles
   //
 
   function irreducible(name, is) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a string
     assert(typeof name === 'string', 'Invalid argument `name` supplied to `irreducible()`');
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a function
     assert(typeof is === 'function', 'Invalid argument `is` supplied to `irreducible()`');
 
     function Irreducible(value, mut, key) {
-
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
       blockNew(this, Irreducible);
-
-      // DEBUG HINT: if the debugger stops here, the first argument is invalid
-      // mouse over the `value` variable to see what's wrong. In `name` there is the name of the type
       assert(is(value), 'Invalid value supplied to `%s` for property `%s`', name, key);
-
       return value;
     }
 
@@ -283,43 +243,29 @@
 
   function struct(props, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a dict of types
-    // mouse over the `props` variable to see what's wrong
-    assert(dict(Str, Type).is(props), 'Invalid argument `props` supplied to `struct()`');
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `struct()`');
-
-    // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || 'struct';
+    assert(dict(Str, Type).is(props), 'Invalid argument `props` = `%s` supplied to `struct` combinator', props);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `struct` combinator', name);
+    name = name || format('{%s}', Object.keys(props).map(function (prop) {
+      return format('%s: %s', prop, getTypeName(props[prop]));
+    }).join(', '));
 
     function Struct(value, mut) {
-
       // makes Struct idempotent
       if (Struct.is(value)) {
         return value;
       }
-
-      // DEBUG HINT: if the debugger stops here, the first argument is invalid
-      // mouse over the `value` variable to see what's wrong. In `name` there is the name of the type
-      assert(Obj.is(value), 'Invalid `%s` supplied to `%s`, expected an `Obj`', value, name);
-
+      assert(Obj.is(value), 'Invalid argument `value` = `%s` supplied to struct type `%s`', value, name);
       // makes `new` optional
       if (!(this instanceof Struct)) {
         return new Struct(value, mut);
       }
-
       for (var k in props) {
         if (props.hasOwnProperty(k)) {
           var expected = props[k];
           var actual = value[k];
-          // DEBUG HINT: if the debugger stops here, the `actual` value supplied to the `expected` type is invalid
-          // mouse over the `actual` and `expected` variables to see what's wrong
           this[k] = expected(actual, mut, k);
         }
       }
-
       if (mut !== true) {
         Object.freeze(this);
       }
@@ -333,18 +279,22 @@
 
     Struct.displayName = name;
 
-    Struct.is = function isStruct(x) {
+    Struct.is = function (x) {
       return x instanceof Struct;
     };
 
-    Struct.update = function updateStruct(instance, spec, value) {
-      return new Struct(update(instance, spec, value));
+    Struct.update = function (instance, spec) {
+      return new Struct(exports.update(instance, spec));
     };
 
-    Struct.extend = function extendStruct(newProps, name) {
-      var newStruct = struct([props].concat(newProps).reduce(mixin, {}), name);
-      mixin(newStruct.prototype, Struct.prototype); // prototypal inheritance
-      return newStruct;
+    Struct.extend = function (arr, name) {
+      arr = [].concat(arr).map(function (x) {
+        return Obj.is(x) ? x : x.meta.props;
+      });
+      arr.unshift(props);
+      var ret = struct(arr.reduce(mixin, {}), name);
+      mixin(ret.prototype, Struct.prototype); // prototypal inheritance
+      return ret;
     };
 
     return Struct;
@@ -352,35 +302,18 @@
 
   function union(types, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a list of types
-    assert(list(Type).is(types), 'Invalid argument `types` supplied to `union()`');
-
+    assert(list(Type).is(types), 'Invalid argument `types` = `%s` supplied to `union` combinator', types);
     var len = types.length;
-
-    // DEBUG HINT: if the debugger stops here, there are too few types (they must be at least two)
-    assert(len >= 2, 'Invalid argument `types` supplied to `union()`');
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `union()`');
-
-    name = name || format('union([%s])', types.map(getName).join(', '));
+    var defaultName = types.map(getTypeName).join(' | ');
+    assert(len >= 2, 'Invalid argument `types` = `%s` supplied to `union` combinator, provide at least two types', defaultName);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `union` combinator', name);
+    name = name || defaultName;
 
     function Union(value, mut) {
-
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
       blockNew(this, Union);
-
-      // DEBUG HINT: if the debugger stops here, you must implement the `dispatch` static method for this type
-      assert(Func.is(Union.dispatch), 'unimplemented %s.dispatch()', name);
-
+      assert(Func.is(Union.dispatch), 'Unimplemented `dispatch()` function for union type `%s`', name);
       var type = Union.dispatch(value);
-
-      // DEBUG HINT: if the debugger stops here, the `dispatch` static method returns no type
-      assert(Type.is(type), '%s.dispatch() returns no type', name);
-
-      // DEBUG HINT: if the debugger stops here, `value` can't be converted to `type`
-      // mouse over the `value` and `type` variables to see what's wrong
+      assert(Type.is(type), 'The `dispatch()` function of union type `%s` returns no type constructor', name);
       return type(value, mut);
     }
 
@@ -392,15 +325,15 @@
 
     Union.displayName = name;
 
-    Union.is = function isUnion(x) {
-      return types.some(function isType(type) {
+    Union.is = function (x) {
+      return types.some(function (type) {
         return type.is(x);
       });
     };
 
     // default dispatch implementation
-    Union.dispatch = function dispatch(x) {
-      for (var i = 0, len = types.length ; i < len ; i++ ) {
+    Union.dispatch = function (x) {
+      for (var i = 0 ; i < len ; i++ ) {
         if (types[i].is(x)) {
           return types[i];
         }
@@ -412,27 +345,16 @@
 
   function maybe(type, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(Type.is(type), 'Invalid argument `type` supplied to `maybe()`');
-
-    // makes the combinator idempotent
-    if (getKind(type) === 'maybe') {
+    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `maybe` combinator', type);
+    // makes the combinator idempotent and handle Any, Nil
+    if (type.meta.kind === 'maybe' || type === Any || type === Nil) {
       return type;
     }
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(Nil.is(name) || Str.is(name), 'Invalid argument `name` supplied to `maybe()`');
-
-    name = name || format('maybe(%s)', getName(type));
+    assert(Nil.is(name) || Str.is(name), 'Invalid argument `name` = `%s` supplied to `maybe` combinator', name);
+    name = name || ('?' + getTypeName(type));
 
     function Maybe(value, mut) {
-
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
       blockNew(this, Maybe);
-
-      // DEBUG HINT: if the debugger stops here, `value` can't be converted to `type`
-      // mouse over the `value` and `type` variables to see what's wrong
       return Nil.is(value) ? null : type(value, mut);
     }
 
@@ -444,7 +366,7 @@
 
     Maybe.displayName = name;
 
-    Maybe.is = function isMaybe(x) {
+    Maybe.is = function (x) {
       return Nil.is(x) || type.is(x);
     };
 
@@ -453,28 +375,14 @@
 
   function enums(map, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a hash
-    // mouse over the `map` variable to see what's wrong
-    assert(Obj.is(map), 'Invalid argument `map` supplied to `enums()`');
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `enums()`');
-
-    name = name || 'enums';
-
-    // cache enums
-    var keys = Object.keys(map);
+    assert(Obj.is(map), 'Invalid argument `map` = `%s` supplied to `enums` combinator', map);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `enums` combinator', name);
+    var keys = Object.keys(map); // cache enums
+    name = name || keys.map(function (k) { return JSON.stringify(k); }).join(' | ');
 
     function Enums(value) {
-
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
       blockNew(this, Enums);
-
-      // DEBUG HINT: if the debugger stops here, the value is not one of the defined enums
-      // mouse over the `value`, `name` and `keys` variables to see what's wrong
-      assert(Enums.is(value), 'Invalid `%s` supplied to `%s`, expected one of %j', value, name, keys);
-
+      assert(Enums.is(value), 'Invalid argument `value` = `%s` supplied to enums type `%s`, expected one of %j', value, name, keys);
       return value;
     }
 
@@ -486,17 +394,17 @@
 
     Enums.displayName = name;
 
-    Enums.is = function isEnums(x) {
+    Enums.is = function (x) {
       return Str.is(x) && map.hasOwnProperty(x);
     };
 
     return Enums;
   }
 
-  enums.of = function enumsOf(keys, name) {
+  enums.of = function (keys, name) {
     keys = Str.is(keys) ? keys.split(' ') : keys;
     var value = {};
-    keys.forEach(function setEnum(k) {
+    keys.forEach(function (k) {
       value[k] = k;
     });
     return enums(value, name);
@@ -504,39 +412,30 @@
 
   function tuple(types, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a list of types
-    assert(list(Type).is(types), 'Invalid argument `types` supplied to `tuple()`');
+    assert(list(Type).is(types), 'Invalid argument `types` = `%s` supplied to `tuple` combinator', types);
+    var len = types.length; // cache types length
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `tuple` combinator', name);
+    name = name || format('[%s]', types.map(getTypeName).join(', '));
 
-    var len = types.length;
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `tuple()`');
-
-    name = name || format('tuple([%s])', types.map(getName).join(', '));
+    function isTuple(x) {
+      return types.every(function (type, i) {
+        return type.is(x[i]);
+      });
+    }
 
     function Tuple(value, mut) {
-
-      // DEBUG HINT: if the debugger stops here, the value is not one of the defined enums
-      // mouse over the `value`, `name` and `len` variables to see what's wrong
-      assert(Arr.is(value) && value.length === len, 'Invalid `%s` supplied to `%s`, expected an `Arr` of length `%s`', value, name, len);
-
+      assert(Arr.is(value) && value.length === len, 'Invalid argument `value` = `%s` supplied to tuple type `%s`, expected an `Arr` of length `%s`', value, name, len);
       var frozen = (mut !== true);
-
       // makes Tuple idempotent
-      if (Tuple.isTuple(value) && Object.isFrozen(value) === frozen) {
+      if (isTuple(value) && Object.isFrozen(value) === frozen) {
         return value;
       }
-
       var arr = [];
       for (var i = 0 ; i < len ; i++) {
         var expected = types[i];
         var actual = value[i];
-        // DEBUG HINT: if the debugger stops here, the `actual` value supplied to the `expected` type is invalid
-        // mouse over the `actual` and `expected` variables to see what's wrong
         arr.push(expected(actual, mut));
       }
-
       if (frozen) {
         Object.freeze(arr);
       }
@@ -552,18 +451,12 @@
 
     Tuple.displayName = name;
 
-    Tuple.isTuple = function isTuple(x) {
-      return types.every(function isType(type, i) {
-        return type.is(x[i]);
-      });
+    Tuple.is = function (x) {
+      return Arr.is(x) && x.length === len && isTuple(x);
     };
 
-    Tuple.is = function isTuple(x) {
-      return Arr.is(x) && x.length === len && Tuple.isTuple(x);
-    };
-
-    Tuple.update = function updateTuple(instance, spec, value) {
-      return Tuple(update(instance, spec, value));
+    Tuple.update = function (instance, spec) {
+      return Tuple(exports.update(instance, spec));
     };
 
     return Tuple;
@@ -571,32 +464,14 @@
 
   function subtype(type, predicate, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(Type.is(type), 'Invalid argument `type` supplied to `subtype()`');
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a function
-    assert(Func.is(predicate), 'Invalid argument `predicate` supplied to `subtype()`');
-
-    // DEBUG HINT: if the debugger stops here, the third argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `subtype()`');
-
-    // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || format('subtype(%s)', getName(type));
-
-    // cache expected value
-    var expected = predicate.__doc__ || format('insert a valid value for %s', predicate.name || 'the subtype');
+    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `subtype` combinator', type);
+    assert(Func.is(predicate), 'Invalid argument `predicate` = `%s` supplied to `subtype` combinator', predicate);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `subtype` combinator', name);
+    name = name || format('{%s | %s}', getTypeName(type), getFunctionName(predicate));
 
     function Subtype(value, mut, key) {
-
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
       blockNew(this, Subtype);
-
-      // DEBUG HINT: if the debugger stops here, the value cannot be converted to the base type
       var x = type(value, mut, key);
-
-      // DEBUG HINT: if the debugger stops here, the value is converted to the base type
-      // but the predicate returns `false`
       assert(predicate(x), 'Invalid value supplied to `%s`, for property `%s`', name, key);
       return x;
     }
@@ -610,12 +485,12 @@
 
     Subtype.displayName = name;
 
-    Subtype.is = function isSubtype(x) {
+    Subtype.is = function (x) {
       return type.is(x) && predicate(x);
     };
 
-    Subtype.update = function updateSubtype(instance, spec, value) {
-      return Subtype(update(instance, spec, value));
+    Subtype.update = function (instance, spec) {
+      return Subtype(exports.update(instance, spec));
     };
 
     return Subtype;
@@ -623,39 +498,26 @@
 
   function list(type, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(Type.is(type), 'Invalid argument `type` supplied to `list()`');
+    assert(Type.is(type), 'Invalid argument `type` = `%s` supplied to `list` combinator', type);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `list` combinator', name);
+    name = name || format('Array<%s>', getTypeName(type));
 
-    // DEBUG HINT: if the debugger stops here, the third argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `list()`');
-
-    // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || format('list(%s)', getName(type));
+    function isList(x) {
+      return x.every(type.is);
+    }
 
     function List(value, mut) {
-
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
-
-      // DEBUG HINT: if the debugger stops here, the value is not one of the defined enums
-      // mouse over the `value` and `name` variables to see what's wrong
-      assert(Arr.is(value), 'Invalid `%s` supplied to `%s`, expected an `Arr`', value, name);
-
+      assert(Arr.is(value), 'Invalid argument `value` = `%s` supplied to list type `%s`', value, name);
       var frozen = (mut !== true);
-
       // makes List idempotent
-      if (List.isList(value) && Object.isFrozen(value) === frozen) {
+      if (isList(value) && Object.isFrozen(value) === frozen) {
         return value;
       }
-
       var arr = [];
       for (var i = 0, len = value.length ; i < len ; i++ ) {
         var actual = value[i];
-        // DEBUG HINT: if the debugger stops here, the `actual` value supplied to the `type` type is invalid
-        // mouse over the `actual` and `type` variables to see what's wrong
         arr.push(type(actual, mut));
       }
-
       if (frozen) {
         Object.freeze(arr);
       }
@@ -670,16 +532,12 @@
 
     List.displayName = name;
 
-    List.isList = function isList(x) {
-      return x.every(type.is);
+    List.is = function (x) {
+      return Arr.is(x) && isList(x);
     };
 
-    List.is = function isList(x) {
-      return Arr.is(x) && List.isList(x);
-    };
-
-    List.update = function updateList(instance, spec, value) {
-      return List(update(instance, spec, value));
+    List.update = function (instance, spec) {
+      return List(exports.update(instance, spec));
     };
 
     return List;
@@ -687,45 +545,35 @@
 
   function dict(domain, codomain, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(Type.is(domain), 'Invalid argument `domain` supplied to `dict()`');
+    assert(Type.is(domain), 'Invalid argument `domain` = `%s` supplied to `dict` combinator', domain);
+    assert(Type.is(codomain), 'Invalid argument `codomain` = `%s` supplied to `dict` combinator', codomain);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `dict` combinator', name);
+    name = name || format('{[key:%s]: %s}', getTypeName(domain), getTypeName(codomain));
 
-    // DEBUG HINT: if the debugger stops here, the second argument is not a type
-    assert(Type.is(codomain), 'Invalid argument `codomain` supplied to `dict()`');
-
-    // DEBUG HINT: if the debugger stops here, the third argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `dict()`');
-
-    // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || format('dict(%s, %s)', getName(domain), getName(codomain));
+    function isDict(x) {
+      for (var k in x) {
+        if (x.hasOwnProperty(k)) {
+          if (!domain.is(k) || !codomain.is(x[k])) { return false; }
+        }
+      }
+      return true;
+    }
 
     function Dict(value, mut) {
-
-      // DEBUG HINT: if the debugger stops here, the value is not an object
-      // mouse over the `value` and `name` variables to see what's wrong
-      assert(Obj.is(value), 'Invalid `%s` supplied to `%s`, expected an `Obj`', value, name);
-
+      assert(Obj.is(value), 'Invalid argument `value` = `%s` supplied to dict type `%s`', value, name);
       var frozen = (mut !== true);
-
       // makes Dict idempotent
-      if (Dict.isDict(value) && Object.isFrozen(value) === frozen) {
+      if (isDict(value) && Object.isFrozen(value) === frozen) {
         return value;
       }
-
       var obj = {};
       for (var k in value) {
         if (value.hasOwnProperty(k)) {
-          // DEBUG HINT: if the debugger stops here, the `k` value supplied to the `domain` type is invalid
-          // mouse over the `k` and `domain` variables to see what's wrong
           k = domain(k);
           var actual = value[k];
-          // DEBUG HINT: if the debugger stops here, the `actual` value supplied to the `codomain` type is invalid
-          // mouse over the `actual` and `codomain` variables to see what's wrong
           obj[k] = codomain(actual, mut);
         }
       }
-
       if (frozen) {
         Object.freeze(obj);
       }
@@ -741,55 +589,37 @@
 
     Dict.displayName = name;
 
-    Dict.isDict = function isDict(x) {
-      for (var k in x) {
-        if (x.hasOwnProperty(k)) {
-          if (!domain.is(k) || !codomain.is(x[k])) { return false; }
-        }
-      }
-      return true;
+    Dict.is = function (x) {
+      return Obj.is(x) && isDict(x);
     };
 
-    Dict.is = function isDict(x) {
-      return Obj.is(x) && Dict.isDict(x);
-    };
-
-
-    Dict.update = function updateDict(instance, spec, value) {
-      return Dict(update(instance, spec, value));
+    Dict.update = function (instance, spec) {
+      return Dict(exports.update(instance, spec));
     };
 
     return Dict;
+  }
+
+  function isInstrumented(f) {
+    return Func.is(f) && Obj.is(f.type);
   }
 
   function func(domain, codomain, name) {
 
     // handle handy syntax for unary functions
     domain = Arr.is(domain) ? domain : [domain];
-
-    // DEBUG HINT: if the debugger stops here, the first argument is not a list of types
-    assert(list(Type).is(domain), 'Invalid argument `domain` supplied to `func()`');
-
-    // DEBUG HINT: if the debugger stops here, the second argument is not a type
-    assert(Type.is(codomain), 'Invalid argument `codomain` supplied to `func()`');
-
-    // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || format('func([%s], %s)', domain.map(getName).join(', '), getName(codomain));
-
-    // cache the domain length
-    var domainLen = domain.length;
+    assert(list(Type).is(domain), 'Invalid argument `domain` = `%s` supplied to `func` combinator', domain);
+    assert(Type.is(codomain), 'Invalid argument `codomain` = `%s` supplied to `func` combinator', codomain);
+    assert(maybe(Str).is(name), 'Invalid argument `name` = `%s` supplied to `func` combinator', name);
+    name = name || format('(%s) => %s', domain.map(getTypeName).join(', '), getTypeName(codomain));
+    var domainLen = domain.length; // cache the domain length
 
     function Func(value) {
-
-      // automatically instrument the function if is not already instrumented
-      if (!func.is(value)) {
-        value = Func.of(value);
+      // automatically instrument the function
+      if (!isInstrumented(value)) {
+        return Func.of(value);
       }
-
-      // DEBUG HINT: if the debugger stops here, the first argument is invalid
-      // mouse over the `value` and `name` variables to see what's wrong
-      assert(Func.is(value), 'Invalid `%s` supplied to `%s`', value, name);
-
+      assert(Func.is(value), 'Invalid argument `value` = `%s` supplied to func type `%s`', value, name);
       return value;
     }
 
@@ -802,18 +632,17 @@
 
     Func.displayName = name;
 
-    Func.is = function isFunc(x) {
-      return func.is(x) &&
-        x.func.domain.length === domain.length &&
-        x.func.domain.every(function isEqual(type, i) {
+    Func.is = function (x) {
+      return isInstrumented(x) &&
+        x.type.domain.length === domainLen &&
+        x.type.domain.every(function (type, i) {
           return type === domain[i];
         }) &&
-        x.func.codomain === codomain;
+        x.type.codomain === codomain;
     };
 
-    Func.of = function funcOf(f) {
+    Func.of = function (f) {
 
-      // DEBUG HINT: if the debugger stops here, f is not a function
       assert(typeof f === 'function');
 
       // makes Func.of idempotent
@@ -822,40 +651,27 @@
       }
 
       function fn() {
-
         var args = slice.call(arguments);
-        var len = Math.min(args.length, domainLen);
-
-        // DEBUG HINT: if the debugger stops here, you provided wrong arguments to the function
-        // mouse over the `args` variable to see what's wrong
-        args = tuple(domain.slice(0, len))(args);
-
+        var len = args.length;
+        var argsType = tuple(domain.slice(0, len));
+        args = argsType(args);
         if (len === domainLen) {
-
           /* jshint validthis: true */
-          var r = f.apply(this, args);
-
-          // DEBUG HINT: if the debugger stops here, the return value of the function is invalid
-          // mouse over the `r` variable to see what's wrong
-          r = codomain(r);
-
-          return r;
-
+          return codomain(f.apply(this, args));
         } else {
-
           var curried = Function.prototype.bind.apply(f, [this].concat(args));
           var newdomain = func(domain.slice(len), codomain);
           return newdomain.of(curried);
-
         }
-
       }
 
-      fn.func = {
+      fn.type = {
         domain: domain,
         codomain: codomain,
         f: f
       };
+
+      fn.displayName = getFunctionName(f);
 
       return fn;
 
@@ -865,27 +681,16 @@
 
   }
 
-  // returns true if x is an instrumented function
-  func.is = function isFunc(f) {
-    return Func.is(f) && Obj.is(f.func);
-  };
-
-  return {
-
-    util: {
-      mixin: mixin,
-      format: format,
-      getName: getName,
-      getKind: getKind,
-      slice: slice,
-      shallowCopy: shallowCopy,
-      update: update
-    },
-
-    options: options,
+  var exports = {
+    format: format,
+    getFunctionName: getFunctionName,
+    getTypeName: getTypeName,
+    mixin: mixin,
+    slice: slice,
+    shallowCopy: shallowCopy,
+    update: update,
     assert: assert,
     fail: fail,
-
     Any: Any,
     Nil: Nil,
     Str: Str,
@@ -898,7 +703,6 @@
     Re: Re,
     Dat: Dat,
     Type: Type,
-
     irreducible: irreducible,
     struct: struct,
     enums: enums,
@@ -910,4 +714,7 @@
     dict: dict,
     func: func
   };
+
+  return exports;
+
 }));